Modify tests to remove external setup dependency
This gets rid of the requirement that the person table exist, and also
removes the need to populate it with the "person:franko" record.

This also changes the test code to use a separate table from our
production data, to avoid polluting the database during testing.

Since we're starting with a fresh table now, I've eliminated the test
that just tried to read back the pre-existing data. It wasn't testing
anything that isn't also being tested in the "insert and get" test
anyway.