Make batch_write_item call conform to public API
The previous way of calling this function worked fine, but it doesn't
follow the type specs defined for the ex_aws API. I'm guessing they may
have changed the API at some point but didn't want to break old code,
and perhaps they left it compatible for that reason. In any case, we
were seeing a Dialyzer warning, and it's better to stick with the
publically documented APIs to avoid any future bugs when we update
ex_aws, so I'm modifying this accordingly.