Fix insert_all return value to conform to Ecto API
According to the Ecto docs, insert_all "returns a tuple containing the
number of entries and any returned results as second element. If the
database does not support RETURNING in UPDATE statements or no return
result was selected, the second element will be `nil`."

Probably would not have noticed this, but Dialyzer compared our
implementation against the Ecto.Adapter callback specs and flagged the
discrepancy.