Fix incorrect range in integer id generation
The hardcoded integer that was previously used to generate this range
was incorrect. It looks like a few digits were dropped from the middle
of the constant. Also, based on the last few digits it looks like the
intended value was 2^128, but we don't actually want to include 2^128 in
our range since you can't fit 2^128 in 128 bits (if we want to generate
128-bit ids, the value at the top of the range should be 2^128 - 1).

This commit fixes those errors, and also improves readability by
generating the correct constant mathematically instead of hardcoding the
39 digits by hand.