support select, fix error in update_ and delete_all, introduce migration
Migration is just introduced here with creat and drop for
tables with a single primary_key. Running the migration
works from the started application (since Hackney needs to be
started), by calling 'Mix.Task.run("ecto.migrate" | "ecto.rollback")'
The primary_key values were assumed to be strings in
update_ and delete_all, this change uses ExAws.Decoder
to decode the DynamoDB attribute/map structure for the
key. Also fixed the case in ecto_dynamo_log when environment
is not set.